12/17일 사이즈,색상 배열 리턴값 ..나오는거수정 ㅠ + 장바구니 총액

+++ src/main/java/com/go/together/Dto/CartDto.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 public class CartDto {
 
     private long cartNumber;
-    private long userNumber;
+    private int userNumber;
     private long productNumber;
     private int cartCount;
